+++ app/src/main/java/kr/sesac/aoao/android/todofolder/ui/TodoFolderActivity.kt
@@ -1,4 +1,4 @@
-package kr.sesac.aoao.android
+package kr.sesac.aoao.android.todofolder.ui
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
